Add files via upload
A kernel (function), on the other hand, is a function of the similarity between two “objects,” colloquially known as feature vectors.

Linear Kernel SVM:

The three classes are separated linearly on the basis of the two best selected attributes -color intensity and proline.

It only works if the data is linearly separable.

It is quite sensitive to outliers.

We can see that the linear kernel works well with the precision of 89%

Precision

0.89

Recall

0.89

F1-Score

0.89

Support

54

Polynomial Kernel SVM:

A low polynomial degree cannot deal with complex datasets, and with high polynomial degree it will create huge number of features, making the model too slow. We can see that the polynomial kernel behaves the best for the given data.

Precision

0.90

Recall

0.89

F1-Score

0.89

Support

54

RBF Kernel SVM:

The RBF kernel is a stationary kernel. It is also known as the “squared exponential” kernel.

can see that RBF Kernel has the least precision among the three kernels.

We

Precision

0.76

Recall

0.74

F1-Score

0.74

Support

54